+++ packages/indiv-loader/compiler/class-property-compiler.ts
@@ -1,0 +1,19 @@
+import types from 'recast/lib/types';
+
+/**
+ * add template in class
+ *
+ * @export
+ * @param {{ templateString: string; classBody: any }} templateInfo
+ * @param {boolean} useTypeScript
+ */
+export function classPropertyCompiler(templateInfo: { templateString: string; classBody: any }, useTypeScript: boolean): void {
+  const typeAnnotation = useTypeScript ? types.builders.typeAnnotation(
+    types.builders.genericTypeAnnotation(types.builders.identifier('string'), null),
+  ) : null;
+  templateInfo.classBody.body.body.push(types.builders.classProperty(
+    types.builders.identifier('template'),
+    types.builders.identifier(`\`${templateInfo.templateString}\``),
+    typeAnnotation,
+  ));
+}